fix bug upload

+++ app/routes/imageRouter.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const router = express.Router();
-const { upload_image } = require("../services/multer");
+const {
+  upload_image
+} = require("../services/multer");
 const fs = require("fs");
 const path = require("path");
 
@@ -13,44 +15,52 @@ router.get("/upload-image", (req, res) => {
 
 router.post("/upload-image", upload_image.single("data-image"), async (req, res) => {
   const extAccept = [".png", ".jpeg", ".jpg"];
-  const extname = path.extname(req.file.originalname).toLowerCase();
+  if (req.file) {
+    const extname = path.extname(req.file.originalname).toLowerCase();
 
-  if (extAccept.includes(extname)) {
-    await fs.renameSync(
-      path.join(__dirname, "../public/img/", req.file.originalname),
-      path.join(__dirname, "../public/img/", "image.png"),
-      (err) => {
-        if (err)
-          throw err;
+    if (extAccept.includes(extname)) {
+      await fs.renameSync(
+        path.join(__dirname, "../public/img/", req.file.originalname),
+        path.join(__dirname, "../public/img/", "image.png"),
+        (err) => {
+          if (err)
+            throw err;
+        }
+      );
+      if (await fs.existsSync(path.join(__dirname, "../../image.png"))) {
+        await fs.unlinkSync(path.join(__dirname, "../../image.png"));
       }
-    );
-    if (await fs.existsSync(path.join(__dirname, "../../image.png"))) {
-      await fs.unlinkSync(path.join(__dirname, "../../image.png"));
+      const path_src = path.join(__dirname, "../public/img/image.png");
+      const path_dst = path.join(__dirname, "../../image.png");
+      await fs.copyFile(path_src, path_dst, (err) => {
+        if (err) {
+          console.log(err);
+        }
+        console.log("Copy file done.");
+      });
+      const hasImage = true;
+      return res.render("upload-image", {
+        hasImage,
+      });
+    } else {
+      if (await fs.existsSync(path.join(__dirname, "../../image.png"))) {
+        await fs.unlinkSync(path.join(__dirname, "../../image.png"));
+      }
+      await fs.unlinkSync(path.join(__dirname, '../public/img/', req.file.originalname));
+
+
+      const notSupport = true;
+      return res.render("upload-image", {
+        notSupport,
+      });
     }
-    const path_src = path.join(__dirname, "../public/img/image.png");
-    const path_dst = path.join(__dirname, "../../image.png");
-    await fs.copyFile(path_src, path_dst, (err) => {
-      if (err) {
-        console.log(err);
-      }
-      console.log("Copy file done.");
-    });
-    const hasImage = true;
-    return res.render("upload-image", {
-      hasImage,
-    });
   } else {
-    if (await fs.existsSync(path.join(__dirname, "../../image.png"))) {
-      await fs.unlinkSync(path.join(__dirname, "../../image.png"));
-    }
-    await fs.unlinkSync(path.join(__dirname, '../public/img/', req.file.originalname));
-
-
-    const notSupport = true;
+    const notFile = true
     return res.render("upload-image", {
-      notSupport,
+      notFile,
     });
   }
+
 });
 
 module.exports = router;
